# What does this PR do?

I recently finished integration your work into SkyRL: https://skyrl.readthedocs.io/en/latest/examples/flash_rl.html 

In my implementation, I've added all patches directly to a custom fork of vLLM: https://github.com/SumanthRH/vllm/tree/flashrl for ease of use. This PR upstreams a change from the SkyRL integration, focusing on vllm 0.9.2. 

I've integrated FlashRL with SkyRL, making it work with the ray backend and additionally two modifications:

### Module-wise memory allocation in patched load weights

 Instead of allocating memory for all the params in `existing_params`, we only allocate memory for the modules corresponding to the parameters in input `weights` list. Since most frameworks gather params layer by layer and issue multiple `load_weights` calls, this is an important optimization. Without this, training OOM'ed for 32B models for online FP8 quant. Along with this, we also modified `process_weights_after_loading`, to only operate on the subset of modules we are interested in. Reference: https://github.com/SumanthRH/vllm/blob/4b04dfcbe8e876e89e4f03859340613d6c766479/vllm/model_executor/layers/patch.py#L659-L741 . Note that **simply modifying the patch here is not enough, since it requires certain guarantees from the training framework** (SkyRL / verl). A naive implementation can break because 
a. vLLM uses QKV fusion: Params for Q, K and V layers should be synced together, and this requires changes in the training framework. My implementation in SkyRL: https://github.com/NovaSky-AI/SkyRL/blob/72467263015fdfe8c8f7ac47040197add22c3418/skyrl-train/skyrl_train/workers/fsdp/fsdp_worker.py#L144-L198
b. vLLM with quantization uses additional weight scaling parameters: It's not enough to just allocate new memory for the parameters in `weights` list, one needs to grab the corresponding modules in vLLM , and then allocate new memory for all parameters in these modules. For example, `self_attn.q_proj.weight` from trainer maps to `self_attn.qkv_proj` module in vLLM. This module will have additional parameters like `weight_scale`. If you don't do this, it will break compatibility with cudagraphs because the memory address for `weight_scale` will change. 


### Improved weight syncing time (140s -> 30s for 32B) for Int8

With v0.9.2, there are some inefficiencies in `process_weights_after_loading`. With int8, FlashRL already handles quantization and mapping new params back to original storage. I found the following while workiing on H100s: 
a. the kenel used for int8 is `CutlassScaledMMLinearKernel` 
b. with the default quant config, in `CutlassScaledMMLinearKernel`,  everything is skipped in the `process_weights_after_loading` except a weight transpose.
     - The way vllm implements this is actually pretty time consuming. This is the main cost in the `process_weights_after_loading` function

We can skip this without any issues because FlashRL's remapping operation is compatible with transpose https://github.com/yaof20/Flash-RL/blob/9d61357b2d060e49610cd55a39bfff422c341382/flash_rl/vllm_patch.py#L579 . For Qwen 32B training on 2 8xH100  nodes with Int8 rollouts, this brought down the time for weight syncing from 140s to 30s, which is lot more reasonable

### Changes in this PR

In this PR, I've only added the second modification. This is because the first one might require changes in `verl` for weight syncing.  I'm curious to get your thoughts @LiyuanLucasLiu 

It's a bit hard to test this rn given my setup is with SkyRL and not verl, so some sanity checks are appreciated. If not, I'll get to it later this weekend. 


### Appendix: Verifying performance for custom `process_weights_after_loading`

I did a micro benchmark for a 14B model where we initialize vLLM engine and then immediately issue a `load_weights` call loading Bf16 weights from `DeepSeek/DeepSeek-R1-Distill-Qwen-14B`. 

```python
###### post init patch in Worker class in vLLM #######
patch_load_weights() # use the custom weights implementation

# load the bf16 model
from transformers import AutoModelForCausalLM
test_model = AutoModelForCausalLM.from_pretrained("DeepSeek/DeepSeek-R1-Distill-Qwen-14B")
full_weights = [(name, param.to(self.device)) for name, param in test_model.named_parameters()]
self.model_runner.model.load_weights(
    weights=full_weights
)
```

Command: 

```bash
lm_eval --model vllm    --model_args pretrained=RedHatAI/DeepSeek-R1-Distill-Qwen-14B-quantized.w8a8,tensor_parallel_size=1,dtype=auto,gpu_memory_utilization=0.8,data_parallel_size=1,enforce_eager=True,distributed_executor_backend=ray --tasks gsm8k     --batch_size auto
```

Baseline: 
a. Performance:
``` bash
vllm (pretrained=RedHatAI/DeepSeek-R1-Distill-Qwen-14B-quantized.w8a8,tensor_parallel_size=1,dtype=auto,gpu_memory_utilization=0.8,data_parallel_size=1,enforce_eager=True), gen_kwargs: (None), limit: None, num_fewshot: None, batch_size: auto
|Tasks|Version|     Filter     |n-shot|  Metric   |   |Value |   |Stderr|
|-----|------:|----------------|-----:|-----------|---|-----:|---|-----:|
|gsm8k|      3|flexible-extract|     5|exact_match|↑  |0.8704|±  |0.0093|
|     |       |strict-match    |     5|exact_match|↑  |0.8681|±  |0.0093|
```

b. E2E time for `patched_load_weights`: 24.41s

skipping `module.process_weights_after_loading` : 
a. Performance:

```bash
vllm (pretrained=RedHatAI/DeepSeek-R1-Distill-Qwen-14B-quantized.w8a8,tensor_parallel_size=1,dtype=auto,gpu_memory_utilization=0.8,data_parallel_size=1,enforce_eager=True), gen_kwargs: (None), limit: None, num_fewshot: None, batch_size: auto
|Tasks|Version|     Filter     |n-shot|  Metric   |   |Value |   |Stderr|
|-----|------:|----------------|-----:|-----------|---|-----:|---|-----:|
|gsm8k|      3|flexible-extract|     5|exact_match|↑  |0.8734|±  |0.0092|
|     |       |strict-match    |     5|exact_match|↑  |0.8719|±  |0.0092|
```

b. E2E time for `patched_load_weights`: 4.66 s

There is a tiny diff in perf but it varies per model - for Qwen 0.5B for example, things matched perfectly. I think it could be due to a lower level difference due to different memory layouts and int8 kernels.

During training, especially for 32B models, since we can't load all the weights at full precision on each rank, the total time for weight syncing is much more (140s)